Savepoint.
Worked on getting the model to work in three steps:
1. Identify items on the board by looking for contours.  This will show
all points but if multiple enemies are touching (or your player is
shooting, they appear as one.)
2. Do a sliding window across it and attempt to classify sprites using a
trained lenet style neural network.
3. Send only the location on a smaller grid (say 80,100) to the final
network. This should allow training much faster.

However, after talking with the ML team at work, they say I should
attempt the RL approach again with a deeper network (similar to a VGG
network) and pass in larger images of the board.

Updated the readme's to reflect my approach and provide more details on
the arduino side.