Update env pre-processing and score delays.
After dumping the state images, it appeared we were getting a lot of
artifacts that could have messed with training.  Also, the score delays
weren't spot on.  By looking at the state images, I was able to dial
those in closer.  It should be playing with rewards closer to the actual
state image and death processing is now accurate.